pyproject.toml
[project]
name = "visu-reborn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,openai,silero]>=1.2.8",
    "pydantic-settings>=2.10.1",
]
